requirements.txt
discord.py==2.0.0
pymongo==4.5.0
mongoengine==0.24.1
python-dotenv==1.0.0

